requirements.txt
openai>=1.3.5
Pillow>=10.1.0

